pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kindle-notes"
version = "0.1.0"
description = "Kindle Scribe notebook extraction and vector render utilities"
readme = ".github/copilot-instructions.md"
requires-python = ">=3.10"
dependencies = [
  "amazon.ion>=0.13.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
vector = [
  "amazon.ion>=0.13.0",
  "Pillow>=10.0.0",
]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
kindle-notes = "kindle_notes.cli:main_cli"

[tool.setuptools]
packages = ["kindle_notes"]
py-modules = ["render_ink"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
